requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
apscheduler==3.10.4
alpaca-py==0.32.0
